// src/app/shared/models/astro-profile.model.ts
export class AstroProfile {
    status: string;
    count: string;
    id: number;
    name: string;
    imgUrl: string;
    expert: string;
    expertise: { id: string, name: string }[];
    experience: string;
    lang: string;
    language: { id: string, name: string }[];
    description: string;
    followerCount: number;
    followerStatus: string;
    isFollow: boolean;
    remainingTime: string;
    talkedTime: string;
    avgRating: string;
    fname: string;
    contact: string;
    email: string;
    isBusy: number;
    isOnline: boolean;
    isActive: string;
    chatStatus: string;

}